requirements.txt
# ============================================================================
# CCTV Camera Analyzer - Production Dependencies
# ============================================================================
# Install: pip install -r requirements.txt
# Python version: 3.8+
# ============================================================================

# ----------------------------------------------------------------------------
# Web Framework
# ----------------------------------------------------------------------------
Flask==3.0.0
Werkzeug==3.0.1

# Production WSGI Server (required for deployment)
gunicorn==21.2.0


# ----------------------------------------------------------------------------
# AI/ML - OpenAI GPT-4o Vision
# ----------------------------------------------------------------------------
openai==1.6.1


# ----------------------------------------------------------------------------
# Google Cloud Storage (Zero Download Mode)
# ----------------------------------------------------------------------------
google-cloud-storage==2.14.0
google-auth==2.25.2


# ----------------------------------------------------------------------------
# Data Processing
# ----------------------------------------------------------------------------
pandas==2.1.4
openpyxl==3.1.2


# ----------------------------------------------------------------------------
# Utilities
# ----------------------------------------------------------------------------
python-dotenv==1.0.0
tqdm==4.66.1
requests==2.31.0


# ============================================================================
# INSTALLATION INSTRUCTIONS
# ============================================================================
#
# 1. Create a virtual environment (recommended):
#    python -m venv venv
#
# 2. Activate the virtual environment:
#    - Windows: venv\Scripts\activate
#    - macOS/Linux: source venv/bin/activate
#
# 3. Install dependencies:
#    pip install -r requirements.txt
#
# 4. Verify installation:
#    python -c "import flask, openai, google.cloud.storage, pandas; print('✅ All dependencies installed!')"
#
# ============================================================================
# OPTIONAL: Development Dependencies (not required for production)
# ============================================================================
# Uncomment below if you need development tools:
#
# pytest==7.4.3              # Testing framework
# pytest-cov==4.1.0          # Code coverage
# black==23.12.1             # Code formatter
# flake8==7.0.0              # Linter
# mypy==1.8.0                # Type checker
# ipython==8.19.0            # Enhanced Python shell
# jupyter==1.0.0             # Jupyter notebooks
#
# ============================================================================

